+++ src/Component/FooterComponents.js/Categories.js
