+++ src/Component/FooterComponents.js/FooterContent.js
